requirements.txt
requests
beautifulsoup4
pillow
gspread
oauth2client